test: extract backend setup helper (WPB-26603)
